feat: add contacts table

+++ src/main/kotlin/br/com/devlhse/kotlinauthservice/domain/repositories/UserRepository.kt
@@ -6,8 +6,8 @@ import br.com.devlhse.kotlinauthservice.domain.model.response.UserResponse
 interface UserRepository {
     fun findAll(): List<UserResponse>
     fun save(user: User)
-    fun findById(id: Long): UserResponse?
+    fun findById(id: Int): UserResponse?
     fun findByEmail(email: String): User?
-    fun update(id: Long, user: User)
-    fun delete(id: Long)
+    fun update(id: Int, user: User)
+    fun delete(id: Int)
 }
